LayoutTests/inspector/canvas/resolveCanvasContext-2d-expected.txt

Tests for the Canvas.resolveCanvasContext command for 2D contexts.


== Running test suite: Canvas.resolveCanvasContext2D
-- Running test case: Canvas.resolveCanvasContext2D.validIdentifier
PASS: Payload should have type "object".
PASS: Payload should have className "CanvasRenderingContext2D".

-- Running test case: Canvas.resolveCanvasContext.invalidIdentifier
PASS: Should produce an error.
Error: Missing canvas for given canvasId
